+++ Structures/InventoryBuilding.cs
@@ -31,7 +31,7 @@ namespace Shared.Structures
 
         public InventoryBuilding(
             HexCell Cell, 
-            int Tribe,
+            byte Tribe,
             byte Level, 
             byte Health, 
             int TroopCount,
